requirements.txt
pandas==2.2.2
numpy==2.0.0
scikit-learn==1.6.1
xgboost==2.0.3
joblib==1.3.2
requests==2.31.0
twilio==8.10.0
python-dotenv==1.0.0
streamlit==1.28.0
plotly==5.17.0
folium==0.15.0
streamlit-folium==0.15.1
schedule==1.2.1
pytz==2023.3